Fixes mistakes from module7-task1 code-review
